services/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "vuniem131104", email = "niemhoalongvu@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "numpy>=2.4.1",
    "pillow>=12.1.0",
    "scipy>=1.17.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "uvicorn>=0.40.0",
]

[project.scripts]
api = "api:main"

[build-system]
requires = ["uv_build>=0.9.4,<0.10.0"]
build-backend = "uv_build"
